fixed bug

+++ stereo_vision/main.h
@@ -12,6 +12,6 @@ void compute_cost(cv::Mat &target, const cv::Mat &imgLeft, const cv::Mat &imgRig
 
 void selectDisparity(cv::Mat &dispLeft, cv::Mat &dispRight, std::vector<cv::Mat> &costVolumeLeft, std::vector<cv::Mat> &costVolumeRight, int scaleDispFactor);
 
-void selectDisparity_v2(cv::Mat &dispLeft, cv::Mat &dispRight, std::vector<cv::Mat> &costVolumeLeft, std::vector<cv::Mat> &costVolumeRight, int scaleDispFactor);
+void selectDisparity_v2(cv::Mat &dispLeft, cv::Mat &dispRight, std::vector<cv::Mat> &costVolumeLeft, std::vector<cv::Mat> &costVolumeRight, float scaleDispFactor);
 
-void refineDisparity(cv::Mat &dispLeft, cv::Mat &dispRight, int scaleDispFactor);
+void refineDisparity(cv::Mat &dispLeft, cv::Mat &dispRight, float scaleDispFactor);
